Implement fetching of titles for the list view and set it dynamically
Add permission to access Network State.
This branch dynalist was created for the dynamic title setting
implementation.Also a footer for the listview has been created, it
will be used to start fetching next set of articles.
The app now checks for internet connectivity and then chooses to
populate the list's titles with hardcoded values or starts to fetch
them accordingly.
The titles in listview of the app is now more readable and in cases of
large titles it doesnt overflow out of the rowview.
A lint error has been introduced now in the adapter of the footer
The app is working as of this commit.

Signed-off-by: Aditya Prasad <adityaprasadiskool@gmail.com>