APP is BROKEN, in the middle of implementing Fetch more articles
The footer's Onclick is only triggering after the first 10 articles
have been fetched.The app crashes after the next 10 titles have been
, during populating of the listview. I belive its because the
string array is set with a size of 10 during the first fetching of
titles.
The units for text size in textview was corrected.
Added the string value to be displayed at the footer during the
time before the first article is fetched.
Also trying to set a onItemClickListener.
Gota Sleep, Jaana.

Signed-off-by: Aditya Prasad <adityaprasadiskool@gmail.com>