Add permission to WRITE_EXTERNAL_STORAGE. Set a better color contrast for the navigation drawer - black.Make class to retrive and save image
In order to save the downloaded image to the sd card, we need the
WRITE_EXTERNAL_STORAGE permission. The greyish transparent background
for the navigation drawer wasnt satisfactory, so I made it black,
which makes the white titles prominent to our eyes.
The new classes are written by [Droidman][1] -
[1]: https://stackoverflow.com/users/1199662/droidman ,
with his permission - "As always, feel free to use this code for any
purposes and modify it in any way." I have modified them to suit my
purpose.
Now once the [primary link][2] has been modified to also contain both
title and image url, the code should fetch the image and store it in the
mobile.while the [secondary link][3] will be used to display the content
when a particular article is selected.
[2]: http://www.motorindiaonline.in/android/?s_i=<start>&e_i=<end>
[3]: http://www.motorindiaonline.in/android/index.php?a_id=<article no>

Signed-off-by: Aditya Prasad <adityaprasadiskool@gmail.com>