Edit README and Implement storage for titles
I have changed the README to reflect the app's features as of now.
I have removed the printing to Log from retriveJSON.java, as its
volume and frequency made debugging other classes difficult.
Also the titles as they are fetched from the server are also stored
via shared preferences, and displayed on the listview if Internet is
not available.

Signed-off-by: Aditya Prasad <adityaprasadiskool@gmail.com>