Show selected article in seperate activity
Finally this is the first working app with all the basic functionality
I have got the selected article to be shown in a seperate activity.
I made a new class retrivearticle just to fetch the article content
this is as the get request to this link returns a JSON object and not
a array of such objects.

The titles and article body seems to contain
some characters which are not recognised by the mobile.
Also when the orientation changes both activites reloads the data
from the server.
well the annoying extra footer bug still persists!! ( I tried spending
 a lot of time trying to figure out how it happens. I only add
the footer in one place! ).
Then there is the strange crash for construction equipment.
awaiting links for lubes,events,dealer corner.

Signed-off-by: Aditya Prasad <adityaprasadiskool@gmail.com>