Implement scrollview, loading animation, Fix crash on quick orientation change, error image size and 2 category links.4
So now the minsdk is set back to jellybean(16), because of the action
bar back button. This can later be brought back to lower levels and
tested for compatibility.
The main article viewing is improved by use of textview inside the
scrollview, which makes scrolling smoother.
I used a webview to show loading animation. I feel it still doesnt
look proffesional.
I changed the image not available picture to motorindia's picture.

Signed-off-by: Aditya Prasad <adityaprasadiskool@gmail.com>