FIX Batteries section
Because the Batteries category was only having a total of 4 articles to
display, its list was never overflowing the screen this caused my
onscrolllistener to keep launching get requests for more articles.
This has been fixed now in case of the Batteries category further
requests for articles will not be sent.
As noted in the code, I do not feel like this category is important
enough to command a section.

Signed-off-by: Aditya Prasad <adityaprasadiskool@gmail.com>